12/25/2014 =============== Gameplay Players can slide along walls & banks of water now Can now aim the range of flamethrower and rocket launcher based on mouse Added homing missile gun =============== UI/HUD implemented translucent Arc heads-up-display of ammo in clip =============== Misc. better console output for debugging, tracing
12/26/2014
=============== Misc.
Reorganized files
=============== Gameplay
Added Nailgun

+++ SydneyShooter/src/sydneyengine/FPSCounter.java
@@ -43,7 +43,7 @@ public class FPSCounter{
 		lastUpdateNanos = newUpdateNanos;
 		counter++;
 		if (cumulativeTimeBetweenUpdatesNanos >= timeBetweenUpdatesNanos){
-			avTimeBetweenUpdatesMillis = (float)((cumulativeTimeBetweenUpdatesNanos)/(counter*1000000f));
+			avTimeBetweenUpdatesMillis = (cumulativeTimeBetweenUpdatesNanos)/(counter*1000000f);
 			freeMemory = Runtime.getRuntime().freeMemory();
 			totalMemory = Runtime.getRuntime().totalMemory();
 			usedMemory = totalMemory - freeMemory;
@@ -59,7 +59,7 @@ public class FPSCounter{
 		return Math.round(getAvTimeBetweenUpdatesMillis());
 	}
 	public float getFPS(){
-		return (float)(getAvTimeBetweenUpdatesMillis() != 0 ? 1000f/getAvTimeBetweenUpdatesMillis() : -1);
+		return getAvTimeBetweenUpdatesMillis() != 0 ? 1000f/getAvTimeBetweenUpdatesMillis() : -1;
 	}
 	public int getFPSRounded(){
 		return Math.round(this.getFPS());
